+++ tests/bunit.web.tests/Asserting/MarkupMatchesAssertExtensionsTest.net5.cs
@@ -1,14 +1,9 @@
 #if NET5_0
 
-using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
 using Bunit.TestAssets.SampleComponents;
 using Xunit;
 
-namespace Bunit
+namespace Bunit.Asserting
 {
 	public partial class MarkupMatchesAssertExtensionsTest : TestContext
 	{